refactor: extract notecard component
